Try to fix the following test case:
EV is charging. Connection to wallbox is interrupted. Charging is stopped in the app while connection is interrupted. Connection is reestablished.
The EV should stop charging now since this is the setting in the app.
Before this commit the EV kept charging when the connection was
reestablished. This should be fixed now.